requirements.txt
Django==5.2.6
djangorestframework==3.16.1
drf-spectacular==0.28.0
psycopg[binary]>=3.1
pytest==8.4.2
pytest-django==4.11.1
ruff==0.13.0
black==24.4.2
mypy==1.18.1
django-stubs==5.2.5
bandit==1.8.6
